chore: move script
